feat: make it responsive
